+++ src/main/kotlin/org/ifinalframework/plugins/aio/mybatis/contributor/MapperXmlCompletionContributor.kt
@@ -1,10 +1,14 @@
 package org.ifinalframework.plugins.aio.mybatis.contributor
 
-import com.intellij.codeInsight.completion.*
+import com.intellij.codeInsight.completion.CompletionParameters
+import com.intellij.codeInsight.completion.CompletionProvider
+import com.intellij.codeInsight.completion.CompletionResultSet
+import com.intellij.codeInsight.completion.CompletionType
 import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.patterns.XmlPatterns
+import com.intellij.psi.PsiEnumConstant
 import com.intellij.psi.xml.XmlToken
 import com.intellij.util.PlatformIcons
 import com.intellij.util.ProcessingContext
@@ -14,11 +18,12 @@ import org.ifinalframework.plugins.aio.mybatis.service.MapperService
 import org.ifinalframework.plugins.aio.mybatis.xml.dom.*
 import org.ifinalframework.plugins.aio.psi.service.DocService
 import org.ifinalframework.plugins.aio.resource.AllIcons
+import org.ifinalframework.plugins.aio.service.PsiService
 import org.ifinalframework.plugins.aio.util.SpiUtil
 import java.util.stream.Stream
 
 /**
- * MapperXmlCompletionContributor
+ * Mapper Xml 自动补全提示
  *
  * @issue 33
  * @author iimik
@@ -31,6 +36,8 @@ class MapperXmlCompletionContributor : AbsMapperCompletionContributor() {
         statementIdCompletion()
         resultMapCompletion()
         resultMapPropertyCompletion()
+        resultMapJdbcTypeCompletion()
+        includeRefidCompletion()
     }
 
 
@@ -148,6 +155,7 @@ class MapperXmlCompletionContributor : AbsMapperCompletionContributor() {
      * <id property=""/>
      * <result property=""/>
      * ```
+     * @see [resultMapJdbcTypeCompletion]
      */
     private fun resultMapPropertyCompletion() {
         val idProperty = XmlPatterns.psiElement().inside(
@@ -200,6 +208,103 @@ class MapperXmlCompletionContributor : AbsMapperCompletionContributor() {
                         }
                     })
             }
+    }
+
+    /**
+     * 自动补全`jdbcType`
+     *
+     * ```xml
+     * <id jdbcType=""/>
+     * <result jdbcType=""/>
+     * ```
+     * @see [resultMapPropertyCompletion]
+     */
+    private fun resultMapJdbcTypeCompletion() {
+        val idProperty = XmlPatterns.psiElement().inside(
+            XmlPatterns.xmlAttribute().withName("jdbcType").inside(XmlPatterns.xmlTag().withName("id"))
+        )
+        val resultProperty = XmlPatterns.psiElement().inside(
+            XmlPatterns.xmlAttribute().withName("jdbcType").inside(XmlPatterns.xmlTag().withName("result"))
+        )
+
+        Stream.of(idProperty, resultProperty)
+            .forEach { place ->
+
+                thisLogger().info("resultMapJdbcTypeCompletion: $place")
+
+                extend(
+                    CompletionType.BASIC,
+                    place,
+                    object : CompletionProvider<CompletionParameters>() {
+                        override fun addCompletions(
+                            parameters: CompletionParameters,
+                            context: ProcessingContext,
+                            result: CompletionResultSet
+                        ) {
+
+                            val project = parameters.position.project
+                            val jdbcTypeClass = project.service<PsiService>().findClass("org.apache.ibatis.type.JdbcType") ?: return
+
+                            val jdbcTypes = jdbcTypeClass.allFields
+                                .filterIsInstance<PsiEnumConstant>()
+                                .map { it.name }
+
+                            jdbcTypes.forEach { jdbcType ->
+                                result.addElement(
+                                    LookupElementBuilder.create(jdbcType)
+                                        .withIcon(AllIcons.Mybatis.JVM)
+                                        .withCaseSensitivity(false)
+                                )
+                            }
+                            result.stopHere()
+
+                        }
+                    })
+            }
+    }
+
+    /**
+     * `include`标签`refid`使用`sql.id`补全提示
+     * ```xml
+     * <include refid="{sqlId}"/>
+     * ```
+     */
+    private fun includeRefidCompletion() {
+        val includeRefid = XmlPatterns.psiElement().inside(
+            XmlPatterns.xmlAttribute().withName("refid").inside(XmlPatterns.xmlTag().withName("include"))
+        )
+
+        thisLogger().info("resultMapPropertyCompletion: $includeRefid")
+
+        extend(
+            CompletionType.BASIC,
+            includeRefid,
+            object : CompletionProvider<CompletionParameters>() {
+                override fun addCompletions(
+                    parameters: CompletionParameters,
+                    context: ProcessingContext,
+                    result: CompletionResultSet
+                ) {
+                    val position = parameters.position
+                    if (position !is XmlToken) return
+                    val domElement = DomUtil.getDomElement(position) ?: return
+                    val mapper = MyBatisUtils.getMapper(domElement)
+
+                    val self = DomUtil.getParentOfType(domElement, Sql::class.java, true)
+
+                    mapper.getSqls()
+                        // 排除自己
+                        .filter { self == null || self.getId().rawText != it.getId().rawText }
+                        .forEach { sql ->
+                            result.addElement(
+                                LookupElementBuilder.create(sql.getId().rawText!!)
+                                    .withIcon(AllIcons.Mybatis.XML)
+                                    .withCaseSensitivity(false)
+                            )
+                        }
+                    result.stopHere()
+                }
+            })
     }